requirements.txt
Django==1.7
FormEncode==1.2.6
IMDbPY==5.0
SQLAlchemy==0.9.7
SQLObject==1.6.0
Tempita==0.5.2
beautifulsoup4==4.3.2
decorator==3.4.0
django-debug-toolbar==1.2.1
lxml==3.4.0
pbr==0.10.0
psycopg2==2.5.4
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.11
wsgiref==0.1.2
